requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
motor==3.3.2
beanie==1.23.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1